Rename zzz to shutdown and add small improvements

+++ .bashrc
@@ -61,14 +61,20 @@ lla()
 }
 alias ls='/bin/ls -F  --color=auto'
 
-zzz()
+shutdown()
 {
+    # $SHUTDOWN_ALLOWED is defined in .bashrc.local
+    if [ "$SHUTDOWN_ALLOWED" != "YES" ]; then
+        echo SHUTDOWN_ALLOWED=YES is not defined
+        return
+    fi
+
     case "$1" in
     "")
         echo "Usage: zzz <minutes>          Shutdown in <minutes> from now"
         echo "       zzz now                Shutdown instandly"
         ;;
-    "now")
+    "n" | "now")
         sudo pkill shutdown
         sudo shutdown -hP now
         ;;
